Testing texture support in application
The application displays dummy textures when the '6' key is pressed on the keyboard. The shader used to render the texture is fairly basic, and does not support lighting as-of this commit. Furthermore, although an overloaded contructor exists, users cannot specify a custom texture yet.

The texture is applied solely on CubeModel objects, and not on Sphere objects.